requirements.txt
# Core framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# Security & Cryptography
cryptography==42.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1

# Data Science & Statistics
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
statsmodels==0.14.2

# Dashboard
streamlit==1.35.0
plotly==5.22.0
reportlab==4.2.0

# Data generation (dev only)
faker==25.2.0

# Logging & monitoring
structlog==24.2.0

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0
